server/requirements.txt


Flask
Flask-PyMongo
Flask-Login
Flask-Cors
Flask-SocketIO
python-dotenv
werkzeug
bcrypt